test: prune redundant refund e2e tests and merge small suites
Cuts ~38 redundant test executions and two per-file boot cycles to speed up CI; every removed test's coverage was verified to survive in a remaining test
