+++ packages/backend/src/services/tx-refunds/create-single-refund.e2e.ts
@@ -88,53 +88,7 @@ describe('Refund Transactions service', () => {
         expect(result.splitId).toEqual(split.id);
       });
 
-      it('successfully creates full refund for split amount', async () => {
-        const account = await helpers.createAccount({ raw: true });
-        const categories = await helpers.getCategoriesList();
-
-        // Create transaction with $30 split
-        const [expenseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[0]!.id,
-            amount: 10000,
-            transactionType: TRANSACTION_TYPES.expense,
-            splits: [{ categoryId: categories[1]!.id, amount: 3000 }],
-          }),
-          raw: true,
-        });
-
-        const transactions = await helpers.getTransactions({
-          raw: true,
-          includeSplits: true,
-        });
-        const split = transactions![0]!.splits![0]!;
-
-        // Create $30 income refund - full split amount
-        const [refundTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[1]!.id,
-            amount: 3000, // Full split amount
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        const result = await helpers.createSingleRefund(
-          {
-            originalTxId: expenseTx.id,
-            refundTxId: refundTx.id,
-            splitId: split.id,
-          },
-          true,
-        );
-
-        expect(result.originalTxId).toEqual(expenseTx.id);
-        expect(result.splitId).toEqual(split.id);
-      });
-
-      it('successfully creates multiple partial refunds for same split', async () => {
+      it('links multiple partial refunds for the same split, and rejects the one that goes over', async () => {
         const account = await helpers.createAccount({ raw: true });
         const categories = await helpers.getCategoriesList();
 
@@ -197,6 +151,28 @@ describe('Refund Transactions service', () => {
           true,
         );
         expect(result2.splitId).toEqual(split.id);
+
+        // Third partial refund: $20 would push the total to $60, past the $50 split
+        const [refundTx3] = await helpers.createTransaction({
+          payload: helpers.buildTransactionPayload({
+            accountId: account.id,
+            categoryId: categories[1]!.id,
+            amount: 2000,
+            transactionType: TRANSACTION_TYPES.income,
+          }),
+          raw: true,
+        });
+
+        const result3 = await helpers.createSingleRefund({
+          originalTxId: expenseTx.id,
+          refundTxId: refundTx3.id,
+          splitId: split.id,
+        });
+
+        expect(result3.statusCode).toEqual(ERROR_CODES.ValidationError);
+        expect(helpers.extractResponse(result3).message).toContain(
+          'Total refund amount cannot be greater than the split amount',
+        );
       });
 
       it('allows refunds on both primary and split independently', async () => {
@@ -310,104 +286,7 @@ describe('Refund Transactions service', () => {
         expect(helpers.extractResponse(result).message).toContain('cannot be greater than the split amount');
       });
 
-      it('fails when total refunds exceed split amount', async () => {
-        const account = await helpers.createAccount({ raw: true });
-        const categories = await helpers.getCategoriesList();
-
-        // Create transaction with $50 split
-        const [expenseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[0]!.id,
-            amount: 10000,
-            transactionType: TRANSACTION_TYPES.expense,
-            splits: [{ categoryId: categories[1]!.id, amount: 5000 }],
-          }),
-          raw: true,
-        });
-
-        const transactions = await helpers.getTransactions({
-          raw: true,
-          includeSplits: true,
-        });
-        const split = transactions![0]!.splits![0]!;
-
-        // First refund: $30
-        const [refundTx1] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[1]!.id,
-            amount: 3000,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        await helpers.createSingleRefund({
-          originalTxId: expenseTx.id,
-          refundTxId: refundTx1.id,
-          splitId: split.id,
-        });
-
-        // Second refund: $30 (total $60 would exceed $50 split)
-        const [refundTx2] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[1]!.id,
-            amount: 3000,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        const result = await helpers.createSingleRefund({
-          originalTxId: expenseTx.id,
-          refundTxId: refundTx2.id,
-          splitId: split.id,
-        });
-
-        expect(result.statusCode).toEqual(ERROR_CODES.ValidationError);
-        expect(helpers.extractResponse(result).message).toContain('cannot be greater than the split amount');
-      });
-
-      it('fails when split does not exist', async () => {
-        const account = await helpers.createAccount({ raw: true });
-        const categories = await helpers.getCategoriesList();
-
-        const [expenseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[0]!.id,
-            amount: 10000,
-            transactionType: TRANSACTION_TYPES.expense,
-          }),
-          raw: true,
-        });
-
-        const [refundTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[0]!.id,
-            amount: 5000,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        // Use a valid UUID format that doesn't exist in the database
-        const fakeUuid = generateRandomRecordId();
-        const result = await helpers.createSingleRefund({
-          originalTxId: expenseTx.id,
-          refundTxId: refundTx.id,
-          splitId: fakeUuid,
-        });
-
-        // Should fail with either NotFoundError (404) or ValidationError (422)
-        expect(result.statusCode).toBeGreaterThanOrEqual(400);
-        expect(result.statusCode).toBeLessThan(500);
-      });
-
-      it('fails when split belongs to different transaction', async () => {
+      it('fails when the split does not exist, and when it belongs to a different transaction', async () => {
         const account = await helpers.createAccount({ raw: true });
         const categories = await helpers.getCategoriesList();
 
@@ -451,54 +330,46 @@ describe('Refund Transactions service', () => {
           raw: true,
         });
 
+        // Valid UUID format that doesn't exist in the database
+        const missingSplitResult = await helpers.createSingleRefund({
+          originalTxId: expenseTx2.id,
+          refundTxId: refundTx.id,
+          splitId: generateRandomRecordId(),
+        });
+
+        expect(missingSplitResult.statusCode).toEqual(ERROR_CODES.NotFoundError);
+        expect(helpers.extractResponse(missingSplitResult).message).toContain('Split not found');
+
         // Try to use split from tx1 with tx2 as original
-        const result = await helpers.createSingleRefund({
+        const foreignSplitResult = await helpers.createSingleRefund({
           originalTxId: expenseTx2.id,
           refundTxId: refundTx.id,
           splitId: splitFromTx1.id,
         });
 
-        expect(result.statusCode).toEqual(ERROR_CODES.ValidationError);
-        expect(helpers.extractResponse(result).message).toContain('does not belong to the original transaction');
+        expect(foreignSplitResult.statusCode).toEqual(ERROR_CODES.ValidationError);
+        expect(helpers.extractResponse(foreignSplitResult).message).toContain(
+          'does not belong to the original transaction',
+        );
       });
 
       it('fails when splitId provided without originalTxId', async () => {
         const account = await helpers.createAccount({ raw: true });
-        const categories = await helpers.getCategoriesList();
-
-        // Create transaction with split to get a valid splitId
-        await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[0]!.id,
-            amount: 10000,
-            transactionType: TRANSACTION_TYPES.expense,
-            splits: [{ categoryId: categories[1]!.id, amount: 3000 }],
-          }),
-          raw: true,
-        });
-
-        const transactions = await helpers.getTransactions({
-          raw: true,
-          includeSplits: true,
-        });
-        const split = transactions![0]!.splits![0]!;
 
         const [refundTx] = await helpers.createTransaction({
           payload: helpers.buildTransactionPayload({
             accountId: account.id,
-            categoryId: categories[1]!.id,
             amount: 2000,
             transactionType: TRANSACTION_TYPES.income,
           }),
           raw: true,
         });
 
-        // Try to create refund with splitId but null originalTxId
+        // The guard rejects before any split lookup, so the id only has to be well-formed.
         const result = await helpers.createSingleRefund({
           originalTxId: null,
           refundTxId: refundTx.id,
-          splitId: split.id,
+          splitId: generateRandomRecordId(),
         });
 
         expect(result.statusCode).toEqual(ERROR_CODES.ValidationError);
@@ -507,65 +378,6 @@ describe('Refund Transactions service', () => {
     });
 
     describe('unlinking and relinking refunds with splits', () => {
-      it('successfully unlinks and relinks refund to same split', async () => {
-        const account = await helpers.createAccount({ raw: true });
-        const categories = await helpers.getCategoriesList();
-
-        const [expenseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[0]!.id,
-            amount: 10000,
-            transactionType: TRANSACTION_TYPES.expense,
-            splits: [{ categoryId: categories[1]!.id, amount: 3000 }],
-          }),
-          raw: true,
-        });
-
-        const transactions = await helpers.getTransactions({
-          raw: true,
-          includeSplits: true,
-        });
-        const split = transactions![0]!.splits![0]!;
-
-        const [refundTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            categoryId: categories[1]!.id,
-            amount: 2000,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        // Create refund link
-        await helpers.createSingleRefund({
-          originalTxId: expenseTx.id,
-          refundTxId: refundTx.id,
-          splitId: split.id,
-        });
-
-        // Unlink
-        const unlinkResult = await helpers.deleteRefund({
-          originalTxId: expenseTx.id,
-          refundTxId: refundTx.id,
-        });
-        expect(unlinkResult.statusCode).toBe(200);
-
-        // Relink
-        const result = await helpers.createSingleRefund(
-          {
-            originalTxId: expenseTx.id,
-            refundTxId: refundTx.id,
-            splitId: split.id,
-          },
-          true,
-        );
-
-        expect(result.originalTxId).toEqual(expenseTx.id);
-        expect(result.splitId).toEqual(split.id);
-      });
-
       it('successfully relinks refund to different split after unlinking', async () => {
         const account = await helpers.createAccount({ raw: true });
         const categories = await helpers.getCategoriesList();
@@ -669,53 +481,6 @@ describe('Refund Transactions service', () => {
       });
 
       it(`successfully creates a refund link between two transactions with different currencies when:
-          – refund amount LESS than base tx amount
-          - refund refAmount LESS than base tx amount
-      `, async () => {
-        const account = await helpers.createAccount({ raw: true });
-        const currencyB = global.MODELS_CURRENCIES!.find((item) => item.code === 'UAH');
-        const accountB = await helpers.createAccount({
-          payload: {
-            ...helpers.buildAccountPayload(),
-            currencyCode: currencyB.code,
-          },
-          raw: true,
-        });
-
-        const [baseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: 100,
-            transactionType: TRANSACTION_TYPES.expense,
-          }),
-          raw: true,
-        });
-
-        const [refundTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: accountB.id,
-            amount: 90,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        const result = await helpers.createSingleRefund(
-          {
-            originalTxId: baseTx.id,
-            refundTxId: refundTx.id,
-          },
-          true,
-        );
-
-        expect(baseTx.currencyCode !== refundTx.currencyCode).toBe(true);
-        expect(baseTx.amount > refundTx.amount).toBe(true);
-        expect(baseTx.refAmount > refundTx.refAmount).toBe(true);
-        expect(result.originalTxId).toEqual(baseTx.id);
-        expect(result.refundTxId).toEqual(refundTx.id);
-      });
-
-      it(`successfully creates a refund link between two transactions with different currencies when:
           – refund amount BIGGER than base tx amount
           - refund refAmount LESS than base tx amount
       `, async () => {
@@ -760,169 +525,6 @@ describe('Refund Transactions service', () => {
         expect(baseTx.refAmount > refundTx.refAmount).toBe(true);
         expect(result.originalTxId).toEqual(baseTx.id);
         expect(result.refundTxId).toEqual(refundTx.id);
-      });
-
-      it('works correctly for cross-account refunds', async () => {
-        const account1 = await helpers.createAccount({ raw: true });
-        const account2 = await helpers.createAccount({
-          payload: helpers.buildAccountPayload({ userId: account1.userId }),
-          raw: true,
-        });
-
-        const [baseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account1.id,
-            amount: 100,
-            transactionType: TRANSACTION_TYPES.expense,
-          }),
-          raw: true,
-        });
-
-        const [refundTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account2.id,
-            amount: 100,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        const result = await helpers.createSingleRefund(
-          {
-            originalTxId: baseTx.id,
-            refundTxId: refundTx.id,
-          },
-          true,
-        );
-
-        expect(baseTx.accountId !== refundTx.accountId).toBe(true);
-        expect(result.originalTxId).toEqual(baseTx.id);
-        expect(result.refundTxId).toEqual(refundTx.id);
-      });
-
-      it.each([
-        [{ originalAmount: 100, refund1: 40, refund2: 60 }], // full refund
-        [{ originalAmount: 100, refund1: 10, refund2: 20 }], // partial refund
-      ])('successfully creates multiple refunds', async ({ originalAmount, refund1, refund2 }) => {
-        const account = await helpers.createAccount({ raw: true });
-
-        const [baseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: originalAmount,
-            transactionType: TRANSACTION_TYPES.expense,
-          }),
-          raw: true,
-        });
-
-        // First partial refund
-        const [refundTx1] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: refund1,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        const result1 = await helpers.createSingleRefund(
-          {
-            originalTxId: baseTx.id,
-            refundTxId: refundTx1.id,
-          },
-          true,
-        );
-
-        expect(result1.originalTxId).toEqual(baseTx.id);
-        expect(result1.refundTxId).toEqual(refundTx1.id);
-
-        // Second partial refund
-        const [refundTx2] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: refund2,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        const result2 = await helpers.createSingleRefund(
-          {
-            originalTxId: baseTx.id,
-            refundTxId: refundTx2.id,
-          },
-          true,
-        );
-
-        expect(result2.originalTxId).toEqual(baseTx.id);
-        expect(result2.refundTxId).toEqual(refundTx2.id);
-      });
-
-      it('successfully creates multiple partial refunds with different currencies', async () => {
-        /**
-         * Create base transaction for which multiple refunds will be added. Then
-         * create two refunds in different currencies.
-         * The point is that second refund is in another currency with lower rate,
-         * so even if we pass bigger `amount`, in fact `refAmount` will be lower,
-         * and this test tests that it works as expected.
-         */
-
-        const amounts = { tx: 1_000, refund_1: 100, refund_2: 1500 };
-
-        const account = await helpers.createAccount({ raw: true });
-        const currencyB = global.MODELS_CURRENCIES!.find((item) => item.code === 'UAH');
-        const accountB = await helpers.createAccount({
-          payload: {
-            ...helpers.buildAccountPayload(),
-            currencyCode: currencyB.code,
-          },
-          raw: true,
-        });
-
-        const [baseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: amounts.tx,
-            transactionType: TRANSACTION_TYPES.expense,
-          }),
-          raw: true,
-        });
-
-        // First partial refund
-        const [refundTx1] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: amounts.refund_1,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        await helpers.createSingleRefund({
-          originalTxId: baseTx.id,
-          refundTxId: refundTx1.id,
-        });
-
-        // Second partial refund with different currency
-        const [refundTx2] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: accountB.id,
-            amount: amounts.refund_2,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        const result = await helpers.createSingleRefund(
-          {
-            originalTxId: baseTx.id,
-            refundTxId: refundTx2.id,
-          },
-          true,
-        );
-
-        expect(result.originalTxId).toEqual(baseTx.id);
-        expect(result.refundTxId).toEqual(refundTx2.id);
       });
 
       it('successfully creates a cross-currency refund whose sum exceeds the original refAmount', async () => {
@@ -1065,82 +667,6 @@ describe('Refund Transactions service', () => {
 
         expect(result.statusCode).toEqual(ERROR_CODES.ValidationError);
         expect(helpers.extractResponse(result).message).toContain('opposite transaction type');
-      });
-
-      it('fails when refund amount is greater than original amount', async () => {
-        const account = await helpers.createAccount({ raw: true });
-
-        const [baseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: 100,
-            transactionType: TRANSACTION_TYPES.expense,
-          }),
-          raw: true,
-        });
-
-        const [refundTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: 150,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        const result = await helpers.createSingleRefund({
-          originalTxId: baseTx.id,
-          refundTxId: refundTx.id,
-        });
-
-        expect(result.statusCode).toEqual(ERROR_CODES.ValidationError);
-        expect(helpers.extractResponse(result).message).toContain('cannot be greater than');
-      });
-
-      it('fails when total refund amount exceeds original transaction amount', async () => {
-        const account = await helpers.createAccount({ raw: true });
-
-        const [baseTx] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: 100,
-            transactionType: TRANSACTION_TYPES.expense,
-          }),
-          raw: true,
-        });
-
-        // First partial refund
-        const [refundTx1] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: 60,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        await helpers.createSingleRefund({
-          originalTxId: baseTx.id,
-          refundTxId: refundTx1.id,
-        });
-
-        // Second partial refund (which would exceed the original amount)
-        const [refundTx2] = await helpers.createTransaction({
-          payload: helpers.buildTransactionPayload({
-            accountId: account.id,
-            amount: 50,
-            transactionType: TRANSACTION_TYPES.income,
-          }),
-          raw: true,
-        });
-
-        const result = await helpers.createSingleRefund({
-          originalTxId: baseTx.id,
-          refundTxId: refundTx2.id,
-        });
-
-        expect(result.statusCode).toEqual(ERROR_CODES.ValidationError);
-        expect(helpers.extractResponse(result).message).toContain('cannot be greater than');
       });
 
       it('fails when trying to refund a transfer transaction', async () => {
